# Syslog parser works with any time format