<!--
Thank you for contributing to Fluentd!
Your commits need to follow DCO: https://probot.github.io/apps/dco/
And please provide the following information to help us make the most of your pull request:
-->

**Which issue(s) this PR fixes**: 
Fixes https://github.com/fluent/fluentd/issues/2856

**What this PR does / why we need it**: 

Setting `time_format`  in syslog hasn't worked well in some situations.
I made a change to work with all formats.  There is no performance degradation.

```rb
require 'benchmark/ips'
require 'fluent/env'
require 'fluent/engine'
require_relative '../lib/fluent/plugin/parser_syslog'

parser = Fluent::Plugin::SyslogParser.new
parser.configure(Fluent::Config::Element.new('ROOT', '', {'keep_time_key' => true}, []))
log = 'Feb 28 12:00:00 192.168.0.1 fluentd[11111]: [error] Syslog test'
parser2 = Fluent::Plugin::SyslogParser.new
parser2.configure(Fluent::Config::Element.new('ROOT', '', {'with_priority' => true, 'keep_time_key' => true}, []))
log2 = '<1>Feb 28 12:00:00 192.168.0.1 fluentd[11111]: [error] Syslog test'

Benchmark.ips do |x|
  x.report "regexp" do
    parser.parse(log) { |t, r| }
  end
  x.report "string" do
    parser.parse_rfc3164(log) { |t, r| }
  end
  x.report "regexp with pri" do
    parser2.parse(log2) { |t, r| }
  end
  x.report "string with pri" do
    parser2.parse_rfc3164(log2) { |t, r| }
  end
end
```

master branch

```
Warming up --------------------------------------
              regexp    22.312k i/100ms
              string    58.311k i/100ms
     regexp with pri    20.555k i/100ms
     string with pri    48.698k i/100ms
Calculating -------------------------------------
              regexp    248.764k (± 4.0%) i/s -      1.249M in   5.031343s
              string    669.950k (± 3.7%) i/s -      3.382M in   5.055510s
     regexp with pri    225.592k (± 3.5%) i/s -      1.131M in   5.017706s
     string with pri    554.281k (± 3.0%) i/s -      2.776M in   5.012618s
```

this branch
```
Warming up --------------------------------------
              regexp    24.435k i/100ms
              string    59.052k i/100ms
     regexp with pri    21.334k i/100ms
     string with pri    48.109k i/100ms
Calculating -------------------------------------
              regexp    262.246k (± 3.3%) i/s -      1.319M in   5.037310s
              string    644.082k (± 2.8%) i/s -      3.248M in   5.046689s
     regexp with pri    225.242k (± 3.5%) i/s -      1.131M in   5.026692s
     string with pri    510.475k (± 7.3%) i/s -      2.550M in   5.026552s
```

**Docs Changes**:

no need

**Release Note**: 

same as the title